# Replace buildout with virtualenv/pip install