--- a/blackbox/bin/hdfs-test
+++ b/blackbox/bin/hdfs-test
@@ -1,0 +1,3 @@
+#!/bin/bash
+DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
+PYTHONPATH=$DIR/../hdfs/src $DIR/../.venv/bin/zope-testrunner --test-path $DIR/../hdfs/src/ "$@"