--- a/blackbox/bin/test
+++ b/blackbox/bin/test
@@ -1,0 +1,3 @@
+#!/bin/bash
+DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
+PYTHONPATH=$DIR/../docs/src $DIR/../.venv/bin/zope-testrunner --test-path $DIR/../docs/src/ "$@"