--- a/blackbox/bootstrap.sh
+++ b/blackbox/bootstrap.sh
@@ -2,7 +2,9 @@
 
 # travis and jenkins have python3.4 installed
 if hash python3.4 2> /dev/null; then
-    python3.4 -m venv .venv
+    python3.4 -m venv .venv --without-pip
+    wget https://bootstrap.pypa.io/get-pip.py
+    .venv/bin/python get-pip.py
 elif hash python3 2> /dev/null; then
     # fallback for dev machines, this should be >= 3.3
     python3 -m venv .venv
@@ -11,4 +13,4 @@ else
     exit 1
 fi
 
-.venv/bin/python bootstrap.py
+.venv/bin/python -m pip install -r requirements.txt