--- a/blackbox/buildout.cfg
+++ b/blackbox/buildout.cfg
@@ -6,7 +6,6 @@ parts = sphinx
         sphinx-cmd
         test
         hdfs-test
-        java_repl
 
 [test]
 recipe = zc.recipe.egg:script
@@ -66,9 +65,3 @@ input = inline:
     ${buildout:bin-directory}/sphinx-build -c docs/ -b text -E docs/clients ${buildout:directory}/docs/clients/out/text
     RESULT+=$?
     exit $RESULT
-
-[java_repl]
-recipe = hexagonit.recipe.download
-url = http://albertlatacz.published.s3.amazonaws.com/javarepl/javarepl.jar
-strip-top-level-dir = true
-download-only = true