--- a/blackbox/buildout.cfg
+++ b/blackbox/buildout.cfg
@@ -31,11 +31,6 @@ initialization=
  sys.argv.extend(['--auto-color',
                   '--test-path', join(base, 'hdfs', 'src'),])
 
-[hadoop]
-recipe = hexagonit.recipe.download
-url = http://www-eu.apache.org/dist/hadoop/common/hadoop-${versions:hadoop}/hadoop-${versions:hadoop}.tar.gz
-strip-top-level-dir = true
-
 [sphinx]
 recipe = zc.recipe.egg:script
 eggs = sphinx