This should speed up the doctests from a clean checkout a bit. Pip can
utilize wheels which is faster than how buildout installs packages.

Before:

    195.64s user 4.90s system 140% cpu 2:22.88 total

With this change:

    134.22s user 3.65s system 235% cpu 58.516 total