@seut 2016-06-16
afaik the outputfile of the `bootstrap` gradle task should be changed to `.venv`